# Ds/src/main/kotlin/org/openapitools/api/AirportsApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package org.openapitools.api

import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.openapitools.model.dto.AirportInfoDto
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Email
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface AirportsApi {

    @Operation(
        tags = ["default",],
        summary = "List all available airports",
        operationId = "airportsLangGet",
        description = """""",
        responses = [
            ApiResponse(responseCode = "200", description = "A list of airports"),
            ApiResponse(responseCode = "400", description = "Invalid language")
        ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/airports/{lang}"],
            produces = ["application/json"]
    )
    fun airportsLangGet(@Parameter(description = "Language code (e.g., en, ru)", required = true, schema = Schema(allowableValues = ["\"en\"", "\"ru\""])) @PathVariable("lang") lang: kotlin.String): ResponseEntity<List<AirportInfoDto>> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }
}