# Ds/src/main/kotlin/org/openapitools/api/ScheduleApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package org.openapitools.api

import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import org.openapitools.model.dto.ScheduledFlightDto
import org.springframework.http.HttpStatus
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated

import javax.validation.Valid

import kotlin.collections.List

@RestController
@Validated
interface ScheduleApi {
    @Operation(
        tags = ["default",],
        summary = "List inbound/outbound flight schedule for an airport",
        operationId = "scheduleAirportCodeLangGet",
        description = """""",
        responses = [
            ApiResponse(responseCode = "200", description = "Schedule list"),
            ApiResponse(responseCode = "400", description = "Invalid parameters provided")
        ]
    )
    @RequestMapping(
        method = [RequestMethod.GET],
        value = ["/schedule/{inbound}/{airport_code}/{lang}"],
        produces = ["application/json"]
    )
    fun scheduleAirportCodeLangGet(@Parameter(description = "", required = true) @PathVariable("inbound") isInbound: kotlin.Boolean, @Parameter(description = "", required = true) @PathVariable("airport_code") airportCode: kotlin.String, @Parameter(description = "Language code (e.g., en, ru)", required = true, schema = Schema(allowableValues = ["\"en\"", "\"ru\""])) @PathVariable("lang") lang: kotlin.String, @Parameter(description = "", schema = Schema(allowableValues = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"])) @Valid @RequestParam(value = "day", required = false) day: kotlin.String?, @Parameter(description = "") @Valid @RequestParam(value = "arrival_time", required = false) arrivalTime: kotlin.String?, @Parameter(description = "") @Valid @RequestParam(value = "flight_number", required = false) flightNumber: kotlin.String?): ResponseEntity<List<ScheduledFlightDto>> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }
}