# Ds/src/main/kotlin/org/openapitools/model/dto/Passenger.kt
package org.openapitools.model.dto

import com.fasterxml.jackson.annotation.JsonProperty
import io.swagger.v3.oas.annotations.media.Schema

/**
 * 
 * @param name 
 */
data class Passenger(

    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("name", required = true) val name: kotlin.String
    ) {

}